Add void type to fix errors with void functions
Void functions used to be uncallable since no return type could be
assigned to them. By adding an explicit void type that is a subtype of
no type and cannot be the type of a variable, the problem is fixed.

In the future, functions should probably have a single type for their
return type, with a special TupleType for multi value return.