Added extensive example file
This file contains a modified version of the tests for the antlr
definition of Solidity. I have changed names to ensure that there are no
naming conflicts, and commented out everything that does not parse.

Things that do not parse:
- Function overloading
- Empty tuple `() = ()`
- Tuple on right side of assignment:
  `var (c,d) = (1, 2);`
- Returning tuples

Things that do not typecheck:
- A fixed size array assigned to a dynamic size array.

There are some warnings, but these are intended for:
- Dead code gives a warning, e.g.
  `continue; break;`
  gives a dead code warning on break.
- Return statements in functions that have no return type defined give a
  warning about a return type mismatch.