Fix variable compilation + operators
Compilation now supports local variables properly. There were mistakes
in the retain-value-at causing it to go into an inifinite loop (hitting
the 100 loops limit). This has now been fixed, so having variables does
no longer cause compilation failure.

Operators requested the NaBL2 AST analysis on a non-ast element, which
fails. This has been resolved by passing the AST analysis as an argument
of the tuple to the op-to-ebc function.